feat(user): 채팅 이름 날짜 보이는 조건 설정
